feat: 10% SIMD speedup
Added unpacking with SIMD for small palette sizes which gives a roughly
10% speedup in program runtime. Also added types for deserializing the
various number array types as bytes and lazily doing endian conversions.

+++ src/distribution.rs
@@ -12,7 +12,7 @@ fn distribution(block_states: &BlockStates) -> Vec<u64> {
 
     let mut distribution = vec![0; block_states.palette.len()];
     for index in block_states.unpack_data() {
-        distribution[index] += 1;
+        distribution[index as usize] += 1;
     }
     distribution
 }
